modified analog resolution to 10bit

+++ uCNC/src/hal/mcus/mcu.h
@@ -163,7 +163,7 @@ extern "C"
  * can be defined either as a function or a macro call
  * */
 #ifndef mcu_get_analog
-	uint8_t mcu_get_analog(uint8_t channel);
+	uint16_t mcu_get_analog(uint8_t channel);
 #endif
 
 /**
